+++ src/components/EducationContainer.css
@@ -1,24 +1,24 @@
-.educationBox {
-  /* Education should be displayed vertically */
-  display: block;
-  flex-direction: column;
-  justify-content: space-around;
-  margin-top: 2rem;
-  margin: 1rem;
+.education-form {
+  width: 500px;
+  margin: 0 auto;
+  border: 1px solid black;
+  padding: 10px;
 }
 
-.education {
-  box-shadow: 0 3px 6px rgba(0, 0, 0, 0.16), 0 3px 6px rgba(0, 0, 0, 0.23);
-  border-radius: 1rem;
-  padding: 0.5rem 0.5rem 0.5rem 1rem;
-  margin: 1rem;
-  text-align: left;
+.education-form input,
+.education-form button {
+  width: 100%;
+  margin-bottom: 10px;
+}
+
+.education-form button {
+  background-color: #000;
+  color: #fff;
+}
+
+.education-form-container {
   display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-}
-
-.dates {
-  font-style: italic;
-  color: "#666";
+  justify-content: center;
+  align-items: center;
+  height: 100vh;
 }
